# docs/persona-ai-mvp.openapi.yaml
openapi: 3.0.3
info:
  title: TalkTo Persona AI MVP API
  version: 0.1.0
  description: |
    Backend MVP API design for Persona AI.

    v1.0 source priority: TalkTo_개발자용_기능명세서_v1.0.md overrides older Notion rows,
    pasted planning notes, and the legacy MVP baseline where they conflict.

    Current implemented surface: legacy Persona Chat MVP plus Archive-first
    subjects, purpose consents, question cards/interactions, recording upload,
    recording completion, recording list/detail, original recording playback,
    payment products/orders/webhooks/entitlements, and queued paid Archive
    analysis jobs, analysis job status APIs, audited admin retry hooks, and
    admin worker transition endpoints for paid analysis jobs, and paid Memories
    status/search plus segment playback/feedback over indexed analysis segments,
    in-app analysis status notifications, and R&D redacted export opt-in/preview
    guardrails, plus Voice Persona application core routes.

    Guardrails: free Archive upload/completion must not create Preview, STT,
    LLM summary, embedding, memory segment, or analysis job work. Preview/sample
    analysis is deferred/disabled by v1.0. POST
    /api/v1/recordings/{recordingId}/preview-analysis exists only as a disabled
    compatibility endpoint returning 409 feature_deferred and must not enqueue
    Preview, free AI, or analysis job work.

    Not yet implemented: automated provider execution for analysis workers,
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/analysis/jobs/{jobId}/retry:
    post:
      tags: [Analysis]
      summary: Retry a retryable paid analysis job
      description: Requeues only owned jobs in failed_retryable state. Completed, failed_terminal, cancelled, queued, and active jobs are rejected; no duplicate active job is created.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      responses:
        '200':
          description: Analysis job requeued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/admin/analysis-jobs:
    get:
      tags: [Admin, Analysis]
      summary: List all analysis jobs
      description: ADMIN and OPS users can list analysis jobs across owners with the same failure code and progress fields exposed to owners.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobStatusFilter'
        - $ref: '#/components/parameters/AnalysisJobRecordingIdFilter'
        - $ref: '#/components/parameters/AnalysisJobSubjectIdFilter'
        - $ref: '#/components/parameters/AnalysisJobLimitFilter'
      responses:
        '200':
          description: All analysis jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/analysis-jobs/{jobId}/retry:
    post:
      tags: [Admin, Analysis]
      summary: Admin retry for a retryable analysis job
      description: ADMIN and OPS users can requeue a failed_retryable job. The existing job row is requeued, duplicate active jobs are not created, and an analysis_job_retry audit log is written.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      responses:
        '200':
          description: Analysis job requeued by admin or ops
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/preprocess:
    patch:
      tags: [Admin, Analysis]
      summary: Mark paid analysis preprocessing
      description: Admin worker transition endpoint. Moves queued/leased work into preprocessing, records the worker id when supplied, and sets a lease expiry used by stale-lease cleanup.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerTransitionRequest'
      responses:
        '200':
          description: Analysis job marked preprocessing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/stt:
    patch:
      tags: [Admin, Analysis]
      summary: Mark STT processing and persist transcript segments
      description: Admin worker transition endpoint. Accepts worker-supplied transcript segments and idempotently upserts them by jobId and segmentIndex.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkSttRequest'
      responses:
        '200':
          description: Analysis job marked STT processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/redaction:
    patch:
      tags: [Admin, Analysis]
      summary: Mark paid analysis redaction gate
      description: Admin worker transition endpoint. Marks the job at the redaction gate without running masking providers.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      responses:
        '200':
          description: Analysis job marked redaction pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/segmenting:
    patch:
      tags: [Admin, Analysis]
      summary: Mark segmenting and persist memory segments
      description: Admin worker transition endpoint. Accepts worker-supplied memory segments and idempotently upserts them by jobId and segmentIndex.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkSegmentingRequest'
      responses:
        '200':
          description: Analysis job marked segmenting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/indexing:
    patch:
      tags: [Admin, Analysis]
      summary: Mark indexing and persist embeddings
      description: Admin worker transition endpoint. Accepts worker-supplied embeddings and idempotently upserts them by jobId and embeddingIndex.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkIndexingRequest'
      responses:
        '200':
          description: Analysis job marked indexing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/completed:
    patch:
      tags: [Admin, Analysis]
      summary: Mark paid analysis completed
      description: Admin worker transition endpoint. Marks the paid analysis job completed and updates recording analysis progress.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      responses:
        '200':
          description: Analysis job marked completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
  /api/v1/admin/analysis/jobs/{jobId}/transitions/failed:
    patch:
      tags: [Admin, Analysis]
      summary: Mark paid analysis failed
      description: Admin worker transition endpoint. Marks the job retryable or terminal according to its retry budget.
      parameters:
        - $ref: '#/components/parameters/AnalysisJobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkFailedRequest'
      responses:
        '200':
          description: Analysis job marked failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
...
  /api/v1/admin/feedback/reviews:
    get:
      tags: [Admin]
      summary: Get feedback review queue
      parameters:
        - in: query
          name: rating
          schema:
            type: string
            enum: [UP, NEUTRAL, DOWN]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Feedback review list
...
  /api/v1/admin/logs/errors:
    get:
      tags: [Admin]
      summary: Get error logs
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [LLM, STT, TTS, AUTH, MEMORY, SYSTEM]
        - in: query
          name: severity
          schema:
            type: string
            enum: [INFO, WARN, ERROR]
      responses:
        '200':
          description: Error logs
  /api/v1/admin/operations/search:
    get:
      tags: [Admin]
      summary: Search sanitized v1 operations resources
      parameters:
        - in: query
          name: q
          schema:
            type: string
            maxLength: 120
        - in: query
          name: status
          schema:
            type: string
            maxLength: 80
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: Sanitized operations search rows
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or token is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
...
    AnalysisJobId:
      in: path
      name: jobId
      required: true
      schema:
        type: string
        format: uuid
...
    AnalysisJobStatusFilter:
      in: query
      name: status
      required: false
      schema:
        type: string
        enum: [queued, leased, preprocessing, stt_processing, redaction_pending, segmenting, embedding, indexing, completed, failed_retryable, failed_terminal, blocked_consent_withdrawn, cancelled]
    AnalysisJobRecordingIdFilter:
      in: query
      name: recordingId
      required: false
      schema:
        type: string
        format: uuid
    AnalysisJobSubjectIdFilter:
      in: query
      name: subjectId
      required: false
      schema:
        type: string
        format: uuid
    AnalysisJobLimitFilter:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: CONFLICT
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
...
    WorkerTransitionRequest:
      type: object
      properties:
        workerId:
          type: string
          maxLength: 120
          example: worker-1
    TranscriptSegmentInput:
      type: object
      required: [segmentIndex, startMs, endMs, transcriptText]
      properties:
        segmentIndex:
          type: integer
          minimum: 0
        startMs:
          type: integer
          minimum: 0
        endMs:
          type: integer
          minimum: 0
        speakerLabel:
          type: string
          maxLength: 120
          example: speaker_1
        transcriptText:
          type: string
          maxLength: 20000
    MarkSttRequest:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TranscriptSegmentInput'
    MemorySegmentInput:
      type: object
      required: [segmentIndex, sourceTranscriptSegmentIds, startMs, endMs, memoryText]
      properties:
        segmentIndex:
          type: integer
          minimum: 0
        sourceTranscriptSegmentIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        startMs:
          type: integer
          minimum: 0
        endMs:
          type: integer
          minimum: 0
        speakerLabel:
          type: string
          maxLength: 120
          example: unknown
        memoryText:
          type: string
          maxLength: 20000
    MarkSegmentingRequest:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MemorySegmentInput'
    EmbeddingInput:
      type: object
      required: [memorySegmentId, embeddingIndex, provider, model, dimensions, embedding]
      properties:
        memorySegmentId:
          type: string
          format: uuid
        embeddingIndex:
          type: integer
          minimum: 0
        provider:
          type: string
          maxLength: 80
          example: synthetic
        model:
          type: string
          maxLength: 120
          example: unit-test
        dimensions:
          type: integer
          minimum: 1
        embedding:
          type: array
          minItems: 1
          items:
            type: number
    MarkIndexingRequest:
      type: object
      required: [embeddings]
      properties:
        embeddings:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/EmbeddingInput'
    MarkFailedRequest:
      type: object
      required: [failureCode]
      properties:
        failureCode:
          type: string
          maxLength: 80
          example: worker_timeout
        failureMessage:
          type: string
          maxLength: 500
          nullable: true
    AnalysisJobResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AnalysisJob'
    AnalysisJobListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisJob'
    AnalysisJobProgress:
      type: object
      properties:
        stage:
          type: string
          enum: [queued, leased, preprocessing, stt_processing, redaction_pending, segmenting, embedding, indexing, completed, failed_retryable, failed_terminal, blocked_consent_withdrawn, cancelled]
        percent:
          type: integer
          minimum: 0
          maximum: 100
        terminal:
          type: boolean
    AnalysisJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        entitlementId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, leased, preprocessing, stt_processing, redaction_pending, segmenting, embedding, indexing, completed, failed_retryable, failed_terminal, blocked_consent_withdrawn, cancelled]
        progress:
          $ref: '#/components/schemas/AnalysisJobProgress'
        retryCount:
          type: integer
        maxRetries:
          type: integer
        failureCode:
          type: string
          nullable: true
        failureMessage:
          type: string
          nullable: true
        statusChangedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        failedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...